chore (general architecture): restructure files into packages that reflect their responsibilities
